requirements.txt
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flask-ujson==1.0.6
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.19.1
MarkupSafe==2.1.3
msgpack==1.0.7
numpy==1.26.2
packaging==23.2
pandas==2.1.3
plotly==5.18.0
psutil==5.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
tenacity==8.2.3
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
